+++ baitap-submit/tainguyen/02-llm-api-params/exercise1.py
@@ -1,0 +1,26 @@
+
+
+import utils
+
+
+def exercise1():
+    while(True):
+        question = utils.require_user_input("Please enter your question: ")
+
+        if question == "e" or question == "q":
+            break
+
+        messages=[{"role": "user", "content": question}]
+
+        stream = utils.client.chat.completions.create(
+            messages=messages,
+            model="gemma2-9b-it",
+            stream=True
+        )
+        print("LLM is thinking...", end="\n\n")
+
+        for chunk in stream:
+            print(chunk.choices[0].delta.content or "", end="")
+
+
+exercise1()
